RFT: moved cli stuff from syn to cli_syn.
